move postLoad code from FieldFillListener.php to ContentFillListener.php
