+++ config/bolt/contenttypes.yaml
@@ -439,18 +439,6 @@ tests:
     listing_records: 10
 
 # This contenttype is here to use for (automated) tests.
-simpletests:
-    name: simpletests
-    singular_name: Simple Test
-    title_format: "{id}: {title}"
-    fields:
-        title:
-            type: text
-            default: "Title of a test contenttype"
-        slug:
-            type: slug
-
-
 complexnestedtype:
     name: complexnestedtype
     singular_name: complexnestedtype
