+++ config/services.yaml
@@ -64,7 +64,6 @@ services:
 
     Bolt\Event\Listener\FieldFillListener:
         tags:
-            - { name: doctrine.event_listener, event: postLoad }
             - { name: doctrine.event_listener, event: preUpdate }
 
     Bolt\Event\Listener\FieldDiscriminatorListener:
