+++ src/Event/Listener/ContentFillListener.php
@@ -5,8 +5,11 @@ declare(strict_types=1);
 namespace Bolt\Event\Listener;
 
 use Bolt\Configuration\Config;
+use Bolt\Configuration\Content\FieldType;
 use Bolt\Entity\Content;
 use Bolt\Entity\Field;
+use Bolt\Entity\Field\CollectionField;
+use Bolt\Entity\Field\SetField;
 use Bolt\Entity\User;
 use Bolt\Enum\Statuses;
 use Bolt\Repository\FieldRepository;
@@ -28,12 +31,16 @@ class ContentFillListener
     /** @var FieldRepository */
     private $fieldRepository;
 
-    public function __construct(Config $config, ContentExtension $contentExtension, UserRepository $users, FieldRepository $fieldRepository)
+    /** @var string */
+    private $defaultLocale;
+
+    public function __construct(Config $config, ContentExtension $contentExtension, UserRepository $users, FieldRepository $fieldRepository, string $defaultLocale)
     {
         $this->config = $config;
         $this->contentExtension = $contentExtension;
         $this->users = $users;
         $this->fieldRepository = $fieldRepository;
+        $this->defaultLocale = $defaultLocale;
     }
 
     public function preUpdate(LifeCycleEventArgs $args): void
@@ -68,6 +75,20 @@ class ContentFillListener
 
         if ($entity instanceof Content) {
             $this->fillContent($entity);
+
+            foreach ($entity->getRawFields() as $rawField) {
+                if ($rawField instanceof Field) {
+                    $this->fillField($rawField);
+                }
+
+                if ($rawField instanceof CollectionField) {
+                    $this->fillCollection($rawField);
+                }
+
+                if ($rawField instanceof SetField) {
+                    $this->fillSet($rawField);
+                }
+            }
         }
     }
 
@@ -155,5 +176,48 @@ class ContentFillListener
         }
 
         return $slug . $separator . '1';
+    }
+
+    public function fillField(Field $field): void
+    {
+        // Fill in the definition of the field
+        $parents = $this->getParents($field);
+        $contentDefinition = $field->getContent()->getDefinition();
+        $field->setDefinition($field->getName(), FieldType::factory($field->getName(), $contentDefinition, $parents));
+        $field->setDefaultLocale($this->defaultLocale);
+
+        $field->setUseDefaultLocale($this->config->get('general/localization')->get('fallback_when_missing'));
+    }
+
+    private function getParents(Field $field): array
+    {
+        $parents = [];
+
+        if ($field->hasParent()) {
+            $parents = $this->getParents($field->getParent());
+            $parents[] = $field->getParent()->getName();
+        }
+
+        return $parents;
+    }
+
+    public function fillSet(SetField $entity): void
+    {
+        $fields = $this->fieldRepository->findAllByParent($entity);
+        $entity->setValue($fields);
+    }
+
+    public function fillCollection(CollectionField $entity): void
+    {
+        $fields = $this->intersectFieldsAndDefinition($this->fieldRepository->findAllByParent($entity), $entity->getDefinition());
+        $entity->setValue($fields);
+    }
+
+    private function intersectFieldsAndDefinition(array $fields, FieldType $definition): array
+    {
+        return
+            collect($fields)->filter(function (Field $field) use ($definition) {
+                return $definition->get('fields') && $definition->get('fields')->has($field->getName());
+            })->values()->toArray();
     }
 }
