+++ src/Event/Listener/FieldFillListener.php
@@ -4,44 +4,22 @@ declare(strict_types=1);
 
 namespace Bolt\Event\Listener;
 
-use Bolt\Configuration\Config;
-use Bolt\Configuration\Content\FieldType;
-use Bolt\Entity\Content;
 use Bolt\Entity\Field;
-use Bolt\Entity\Field\CollectionField;
 use Bolt\Entity\Field\RawPersistable;
-use Bolt\Entity\Field\SetField;
 use Bolt\Entity\FieldInterface;
 use Bolt\Entity\FieldTranslation;
-use Bolt\Repository\FieldRepository;
 use Bolt\Utils\Sanitiser;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Twig\Markup;
 
 class FieldFillListener
 {
-    /** @var FieldRepository */
-    private $fields;
-
-    /** @var ContentFillListener */
-    private $cfl;
-
     /** @var Sanitiser */
     private $sanitiser;
 
-    /** @var string */
-    private $defaultLocale;
-
-    /** @var Config */
-    private $config;
-
-    public function __construct(FieldRepository $fields, ContentFillListener $cfl, Sanitiser $sanitiser, string $defaultLocale, Config $config)
+    public function __construct(Sanitiser $sanitiser)
     {
-        $this->fields = $fields;
-        $this->cfl = $cfl;
         $this->sanitiser = $sanitiser;
-        $this->defaultLocale = $defaultLocale;
-        $this->config = $config;
     }
 
     public function preUpdate(LifecycleEventArgs $args): void
@@ -88,107 +66,5 @@ class FieldFillListener
     public static function trimZeroWidthWhitespace(string $string): string
     {
         return preg_replace('/([{}])[\x{200B}-\x{200D}\x{FEFF}]([{}])/u', '$1$2', $string);
-    }
-
-    public function postLoad(LifecycleEventArgs $args): void
-    {
-        $entity = $args->getEntity();
-
-//        Disable indivitual 'filling' of fields
-//
-//        if ($entity instanceof Field) {
-//            $this->fillField($entity);
-//        }
-//
-//        if ($entity instanceof CollectionField) {
-//            $this->fillCollection($entity);
-//        }
-//
-//        if ($entity instanceof SetField) {
-//            $this->fillSet($entity);
-//        }
-
-//        instead only fill once, when content entity has been loaded
-//        note this functionality probably should be moved to the ContentFillListener, then the
-//        FieldFillListener can simply be removed. (However, note that the config-checker package will complain!)
-        if ($entity instanceof Content) {
-            foreach ($entity->getRawFields() as $rawField) {
-                if ($rawField instanceof Field) {
-                    $this->fillField($rawField);
-                }
-
-                if ($rawField instanceof CollectionField) {
-                    $this->fillCollection($rawField);
-                }
-
-                if ($rawField instanceof SetField) {
-                    $this->fillSet($rawField);
-                }
-            }
-        }
-
-//
-//        Comment the 3 ifs above, and uncomment code below to fix initialization of selects
-//
-
-//        if ($entity instanceof Content) {
-//            foreach ($entity->getRawFields() as $rawField) {
-//                if ($rawField instanceof Field) {
-//                    $this->fillField($rawField);
-//                }
-//
-//                if ($rawField instanceof CollectionField) {
-//                    $this->fillCollection($rawField);
-//                }
-//
-//                if ($rawField instanceof SetField) {
-//                    $this->fillSet($rawField);
-//                }
-//            }
-//        }
-    }
-
-    public function fillField(Field $field): void
-    {
-        // Fill in the definition of the field
-        $parents = $this->getParents($field);
-        $this->cfl->fillContent($field->getContent());
-        $contentDefinition = $field->getContent()->getDefinition();
-        $field->setDefinition($field->getName(), FieldType::factory($field->getName(), $contentDefinition, $parents));
-        $field->setDefaultLocale($this->defaultLocale);
-
-        $field->setUseDefaultLocale($this->config->get('general/localization')->get('fallback_when_missing'));
-    }
-
-    private function getParents(Field $field): array
-    {
-        $parents = [];
-
-        if ($field->hasParent()) {
-            $parents = $this->getParents($field->getParent());
-            $parents[] = $field->getParent()->getName();
-        }
-
-        return $parents;
-    }
-
-    public function fillSet(SetField $entity): void
-    {
-        $fields = $this->fields->findAllByParent($entity);
-        $entity->setValue($fields);
-    }
-
-    public function fillCollection(CollectionField $entity): void
-    {
-        $fields = $this->intersectFieldsAndDefinition($this->fields->findAllByParent($entity), $entity->getDefinition());
-        $entity->setValue($fields);
-    }
-
-    private function intersectFieldsAndDefinition(array $fields, FieldType $definition): array
-    {
-        return
-            collect($fields)->filter(function (Field $field) use ($definition) {
-                return $definition->get('fields') && $definition->get('fields')->has($field->getName());
-            })->values()->toArray();
     }
 }
