+++ tests/php/Configuration/Parser/ContentTypesParserTest.php
@@ -70,7 +70,7 @@ class ContentTypesParserTest extends ParserTestBase
         $contentTypesParser = new ContentTypesParser($this->getProjectDir(), $generalParser->parse(), self::DEFAULT_LOCALE, self::ALLOWED_LOCALES);
         $config = $contentTypesParser->parse();
 
-        $this->assertCount(7, $config);
+        $this->assertCount(8, $config);
 
         $this->assertArrayHasKey('homepage', $config);
         $this->assertCount(self::AMOUNT_OF_ATTRIBUTES_IN_CONTENT_TYPE, $config['homepage']);
